Add a hook for crashreporting
Add a hook settable via cygwin_internal() which is invoked on a fatal signal and
has access to the EXCEPTION_POINTERS for the exception.

Also, slightly rearrange things in signal_exit(), which has the side-effect of
making the DLL try to break to a debugger on a fatal signal, even if core file
are turned off with 'ulimit -c 0'.  I'm not sure why we would want it the
current way anyhow.